Create Solution.cpp
This is the solution for the Assignment question given to me.
The maxProfit function takes a vector of prices as input and returns the maximum profit that can be achieved with at most two transactions. The function works as follows:

First, it initializes two arrays, leftProfit and rightProfit, of length n, where n is the length of the input vector prices. The leftProfit[i] (resp. rightProfit[i]) represents the maximum profit that can be obtained by making a single transaction on the left (resp. right) of day i.
Then, it computes the leftProfit array by iterating over the input vector from left to right. At each position i, it updates the leftMin variable to the minimum price seen so far and sets leftProfit[i] to the maximum of the previous leftProfit[i-1] and the profit that can be obtained by buying at the leftMin price and selling at the current prices[i].
Similarly, it computes the rightProfit array by iterating over the input vector from right to left. At each position i, it updates the rightMax variable to the maximum price seen so far and sets rightProfit[i] to the maximum of the previous rightProfit[i+1] and the profit that can be obtained by buying at the current prices[i] and selling at the rightMax price.
Finally, it computes the maximum profit that can be obtained by making two transactions by summing up the leftProfit and rightProfit arrays for each position i and keeping track of the maximum value seen so far.